### 2026-05-07 #10558 @editorialbot
[PRE REVIEW]: Potamides
**Submitting author:** @nstarman (<a href="http://orcid.org/0000-0003-3954-3291">Nathaniel Starkman</a>)
**Repository:** https://github.com/xggs-dev/potamides
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
