### 2026-05-08 ba317e6
docs(paper): revise JOSS paper text and bibliography
Update paper.md with updated title, expanded summary, statement of need, field context, software design, and research impact sections. Add AI usage disclosure. Update bibliography to standard BibTeX format.

Signed-off-by: nstarman <nstarman@users.noreply.github.com>
