+++ paper/paper.md
@@ -1,9 +1,10 @@
 ---
-title: "Potamides: A Python package for stream curvature analysis"
+title: "Potamides: JAX tools for curvature-based inference from stellar streams"
 tags:
   - Python
   - astronomy
-  - stellar stream
+  - stellar streams
+  - galactic dynamics
 authors:
   - name: Sirui Wu
     orcid: 0009-0003-4675-3622
@@ -21,7 +22,6 @@ authors:
     orcid: 0000-0003-0256-5446
     corresponding: true
     affiliation: 1
-
 affiliations:
   - name: Niels Bohr Institute, University of Copenhagen, Denmark
     index: 1
@@ -40,74 +40,124 @@ bibliography: paper.bib
 
 # Summary
 
-Potamides is a Python package for inferring galactic gravitational potentials
-from the geometry of stellar streams. Stellar streams—elongated structures
-formed by stars tidally stripped from globular clusters or dwarf galaxies—trace
-the gravitational field of their host galaxy. Their local curvature contains
-information about the underlying mass distribution, enabling direct inference of
-dark matter halo shape and orientation, baryonic components such as the disk,
-and the position of the galactic center.
+`Potamides` is a Python package for inferring the mass distribution of galaxies
+from the projected shapes of stellar streams in imaging data. Stellar streams
+are elongated structures produced when star clusters or dwarf galaxies are
+tidally disrupted by their host. Because their projected tracks carry
+information about the host's gravitational field, the local curvature of a
+stream can constrain the underlying potential.
 
-`Potamides` implements the curvature–gravitational-acceleration alignment
-likelihood framework introduced by [@Nibauer:2023]. The package couples
-JAX-accelerated spline representations of stream tracks with fast evaluations of
-gravitational accelerations from flexible halo and disk potentials. This
-supports an end-to-end workflow: constructing smooth stream tracks from observed
-positions; computing tangents, principal normals, and scalar curvature;
-evaluating gravitational accelerations under candidate potentials; and combining
-segment-wise likelihoods across multiple streams using density-based weighting.
-
-The package integrates with `unxt` [@unxt] for JAX-compatible units and with
-`galax` [@galax] for gravitational potential evaluation. JAX’s just-in-time
-(JIT) compilation and vectorization enable efficient evaluation of
-curvature-based likelihoods, which is essential for Bayesian inference and
-parameter exploration with modern samplers. By focusing on stream curvature
-rather than explicit orbit integration, `Potamides` provides a complementary
-framework for gravitational potential inference in galactic dynamics. The
-package includes comprehensive tests and documented examples, making
-high-performance, curvature-based inference accessible to the astronomical
-community.
+The package implements and extends the curvature-based likelihood framework of
+[@Nibauer:2023]. Rather than generating a full dynamical realization of a
+stellar stream for each trial model, `Potamides` represents observed stream
+tracks with JAX-based splines. It evaluates gravitational accelerations in
+candidate potentials and compares them directly to the local stream geometry.
+This provides a lower-cost inference workflow that complements traditional
+forward-modeling approaches. `Potamides` supports the complete analysis pipeline
+for a galaxy, from annotating stream ridge-lines to evaluating likelihoods
+across many potential models.
 
 # Statement of need
 
-Constraining the shape and structure of dark matter halos is central to
-understanding galaxy formation and testing cosmological models. Stellar
-streams—tidal debris from disrupted satellites such as dwarf galaxies or
-globular clusters—serve as sensitive tracers of the galactic gravitational
-potential because their morphology encodes the host halo’s properties
-[@Bonaca:2014]. The curvature-based inference method introduced by
-[@Nibauer:2023] provides a novel approach by comparing the local curvature of
-observed streams with predicted gravitational accelerations, enabling robust
-constraints on halo flattening and orientation. Until recently, however, this
[…]
