+++ paper/paper.bib
@@ -1,84 +1,106 @@
-@ARTICLE{Mateu:2023,
-       author = {{Mateu}, Cecilia},
-        title = "{galstreams: A library of Milky Way stellar stream footprints and tracks}",
-      journal = {\mnras},
-     keywords = {Astronomical databases: catalogues, Software: public release, Galaxy: halo, Galaxy: structure, Astrophysics - Astrophysics of Galaxies},
-         year = 2023,
-        month = apr,
-       volume = {520},
-       number = {4},
-        pages = {5225-5258},
-          doi = {10.1093/mnras/stad321},
-archivePrefix = {arXiv},
-       eprint = {2204.10326},
- primaryClass = {astro-ph.GA},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2023MNRAS.520.5225M},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+@article{Mateu:2023,
+  author = {Mateu, Cecilia},
+  title = {galstreams: A library of Milky Way stellar stream footprints and tracks},
+  journal = {Monthly Notices of the Royal Astronomical Society},
+  year = {2023},
+  volume = {520},
+  number = {4},
+  pages = {5225--5258},
+  doi = {10.1093/mnras/stad321}
 }
 
-@ARTICLE{Nibauer:2023,
-       author = {{Nibauer}, Jacob and {Bonaca}, Ana and {Johnston}, Kathryn V.},
-        title = "{Constraining the Gravitational Potential from the Projected Morphology of Extragalactic Tidal Streams}",
-      journal = {\apj},
-     keywords = {Stellar streams, Galaxy stellar halos, Dark matter, Tidal tails, 2166, 598, 353, 1701, Astrophysics - Astrophysics of Galaxies},
-         year = 2023,
-        month = sep,
-       volume = {954},
-       number = {2},
-          eid = {195},
-        pages = {195},
-          doi = {10.3847/1538-4357/ace9bc},
-archivePrefix = {arXiv},
-       eprint = {2303.17406},
- primaryClass = {astro-ph.GA},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2023ApJ...954..195N},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+@article{Nibauer:2023,
+  author = {Nibauer, Jacob and Bonaca, Ana and Johnston, Kathryn V.},
+  title = {Constraining the Gravitational Potential from the Projected Morphology of Extragalactic Tidal Streams},
+  journal = {The Astrophysical Journal},
+  year = {2023},
+  volume = {954},
+  number = {2},
+  pages = {195},
+  doi = {10.3847/1538-4357/ace9bc}
 }
-
 
 @article{unxt,
-  author    = {Starkman, Nathaniel and Price-Whelan, Adrian M. and Nibauer, Jake},
-  title     = {{unxt}: A {Python} package for unit-aware computing with {JAX}},
-  journal   = {Journal of Open Source Software},
-  year      = {2025},
-  doi       = {10.21105/joss.07771},
-  url       = {https://github.com/GalacticDynamics/unxt},
+  author = {Starkman, Nathaniel and Price-Whelan, Adrian M. and Nibauer, Jacob},
+  title = {{unxt}: A Python package for unit-aware computing with {JAX}},
+  journal = {Journal of Open Source Software},
+  year = {2025},
+  volume = {10},
+  number = {106},
+  pages = {7771},
+  doi = {10.21105/joss.07771},
+  url = {https://github.com/GalacticDynamics/unxt}
 }
 
 @software{galax,
[…]
