### 2026-06-16 #10712 @editorialbot
[REVIEW]: Potamides: JAX tools for curvature-based inference from stellar streams
**Submitting author:** @nstarman (<a href="http://orcid.org/0000-0003-3954-3291">Nathaniel Starkman</a>)
**Repository:** https://github.com/xggs-dev/potamides
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
