### 2026-06-22 22c9a51
Constrain quax to 0.3.5+
