### 2026-06-27 e9f05cb
Add Contributor Covenant Code of Conduct
Added Contributor Covenant Code of Conduct to establish community standards and expectations.

Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
