### 2026-06-27 49a43aa
Fix prettier formatting in CODE_OF_CONDUCT.md
