### 2026-06-27 00824fe
Add Contributor Covenant Code of Conduct (#100)
* Add Contributor Covenant Code of Conduct

Added Contributor Covenant Code of Conduct to establish community standards and expectations.

Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
Co-authored-by: Copilot Autofix powered by AI <175728472+Copilot@users.noreply.github.com>
Co-authored-by: copilot-swe-agent[bot] <198982749+Copilot@users.noreply.github.com>
