### 2026-06-27 5db92ed
docs(paper): refine Nibauer citation
Signed-off-by: nstarman <nstarman@users.noreply.github.com>
