### 2026-06-29 6ae8d11
docs(paper): address JOSS review comment (#101)
* docs(paper): refine Nibauer citation
* Potential fix for pull request finding

Signed-off-by: nstarman <nstarman@users.noreply.github.com>
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
Co-authored-by: Copilot Autofix powered by AI <175728472+Copilot@users.noreply.github.com>
