+++ .github/workflows/cd.yml
@@ -29,7 +29,7 @@ jobs:
       attestations: write
 
     steps:
-      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v5.5.0
+      - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v5.5.0
         with:
           # Required for hatch-vcs to determine version from git tags.
           # Can be removed if switching to hardcoded versions.
