### 2026-07-13 4dae289
docs: add CONTRIBUTING guide and link from README
Introduce a CONTRIBUTING.md covering ways to contribute, development setup with uv and nox, commit conventions, and CI expectations. Add a Contributing section to the README pointing users to the new guide and standard GitHub workflows.

Signed-off-by: nstarman <nstarman@users.noreply.github.com>
