### 2026-07-13 c0a193d
docs(paper): add Hoffman & Gelman NUTS reference and tidy bib entries
Add the Hoffman & Gelman (2011) citation for NUTS, update the paper.md reference to point to it, and alphabetize/reformat the bib entries for consistent field ordering.

Signed-off-by: nstarman <nstarman@users.noreply.github.com>
