+++ paper/paper.bib
@@ -1,106 +1,123 @@
-@article{Mateu:2023,
-  author = {Mateu, Cecilia},
-  title = {galstreams: A library of Milky Way stellar stream footprints and tracks},
-  journal = {Monthly Notices of the Royal Astronomical Society},
-  year = {2023},
-  volume = {520},
-  number = {4},
-  pages = {5225--5258},
-  doi = {10.1093/mnras/stad321}
-}
-
-@article{Nibauer:2023,
-  author = {Nibauer, Jacob and Bonaca, Ana and Johnston, Kathryn V.},
-  title = {Constraining the Gravitational Potential from the Projected Morphology of Extragalactic Tidal Streams},
+@article{Bonaca:2014,
+  author  = {Bonaca, Ana and Geha, Marla and K{\"u}pper, Andreas H. W. and Diemand, J{\"u}rg and Johnston, Kathryn V. and Hogg, David W.},
+  doi     = {10.1088/0004-637X/795/1/94},
   journal = {The Astrophysical Journal},
-  year = {2023},
-  volume = {954},
-  number = {2},
-  pages = {195},
-  doi = {10.3847/1538-4357/ace9bc}
+  number  = {1},
+  pages   = {94},
+  title   = {Milky Way Mass and Potential Recovery Using Tidal Streams in a Realistic Halo},
+  volume  = {795},
+  year    = {2014}
 }
 
-@article{unxt,
-  author = {Starkman, Nathaniel and Price-Whelan, Adrian M. and Nibauer, Jacob},
-  title = {{unxt}: A Python package for unit-aware computing with {JAX}},
-  journal = {Journal of Open Source Software},
-  year = {2025},
-  volume = {10},
-  number = {106},
-  pages = {7771},
-  doi = {10.21105/joss.07771},
-  url = {https://github.com/GalacticDynamics/unxt}
+@article{Bovy2015,
+  author  = {Bovy, Jo},
+  doi     = {10.1088/0067-0049/216/2/29},
+  journal = {The Astrophysical Journal Supplement Series},
+  number  = {2},
+  pages   = {29},
+  title   = {galpy: A Python Library for Galactic Dynamics},
+  volume  = {216},
+  year    = {2015}
 }
 
 @software{galax,
-  author = {Starkman, Nathaniel and Price-Whelan, Adrian M. and Nibauer, Jacob},
-  title = {{GalacticDynamics/galax}: v0.0.2},
-  year = {2024},
-  month = jun,
+  author    = {Starkman, Nathaniel and Price-Whelan, Adrian M. and Nibauer, Jacob},
+  doi       = {10.5281/zenodo.11553485},
+  month     = jun,
   publisher = {Zenodo},
-  version = {v0.0.2},
-  doi = {10.5281/zenodo.11553485},
-  url = {https://github.com/GalacticDynamics/galax}
+  title     = {{GalacticDynamics/galax}: v0.0.2},
+  url       = {https://github.com/GalacticDynamics/galax},
+  version   = {v0.0.2},
+  year      = {2024}
+}
+
+@article{Hoffman+Gelman:2011,
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2011arXiv1111.4246H},
+  archiveprefix = {arXiv},
+  author        = {{Hoffman}, Matthew D. and {Gelman}, Andrew},
+  doi           = {10.48550/arXiv.1111.4246},
+  eid           = {arXiv:1111.4246},
+  eprint        = {1111.4246},
+  journal       = {arXiv e-prints},
+  keywords      = {Statistics - Computation, Computer Science - Machine Learning},
+  month         = nov,
[…]
