### 2026-07-13 4223b6e
docs: add contributing guide and readme, and fix paper citation (#104)
* docs: add CONTRIBUTING guide and link from README

Introduce a CONTRIBUTING.md covering ways to contribute, development setup with uv and nox, commit conventions, and CI expectations. Add a Contributing section to the README pointing users to the new guide and standard GitHub workflows.

* docs(paper): add Hoffman & Gelman NUTS reference and tidy bib entries

Add the Hoffman & Gelman (2011) citation for NUTS, update the paper.md reference to point to it, and alphabetize/reformat the bib entries for consistent field orderi
[…]
