### 2026-08-16 #10712 @TobiBu
@nstarman what is the status on the installation fix? This might be the only thing blocking to move forward...
