### 2026-08-17 37bd316
build(deps): use galax v0.0.3 from PyPI
Replace the git-main pin on galax with the PyPI release `galax>=0.0.3`.
galax 0.0.3 requires Python >=3.12, so bump `requires-python` accordingly
(classifiers, mypy, pylint, CI matrix, README). Drop the `quax>=0.3.5` uv
constraint: galax 0.0.3 already requires quax>=0.4.3 and resolution is
unchanged without it.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
