+++ .github/workflows/ci.yml
@@ -41,7 +41,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.11", "3.13"]
+        python-version: ["3.12", "3.13"]
         runs-on: [ubuntu-latest, windows-latest, macos-14]
 
     steps:
