### 2026-08-17 b41f2bc
build(deps): bump build-and-inspect-python-package to v3.0.1 (#108)
hatchling now emits `Metadata-Version: 2.5`, which the Twine bundled in
baipp v2.18.0 rejects: `twine check --strict` fails with "'2.5' is not a
valid metadata version". baipp v3.0.0 ships Twine 7, which supports
metadata 2.5.

Also correct the stale version comment on the `actions/checkout` pin;
SHA 3d3c42e is v7.0.1, not v5.5.0.

Co-authored-by: Claude Opus 5 <noreply@anthropic.com>
