### 2026-08-17 544cbc1
style: ignore ruff UP040 for jaxtyping type aliases
Bumping `requires-python` to >=3.12 makes ruff enforce UP040, flagging all
27 `X: TypeAlias = ...` jaxtyping shape aliases. Converting them to PEP 695
`type` statements is not viable here: the pinned mypy (v1.15.0 in
pre-commit, 1.16.1 locally) cannot parse `type` statements at all, so the
conversion breaks type checking outright.

Ignore the rule instead, alongside the existing jaxtyping-related F722 and
F821 ignores.

Co-Authored-By: Claude Opus 5 <noreply@anthropic.com>
