### 2026-08-17 #10712 @nstarman
I just cut a new release of https://github.com/xggs-dev/potamides that implements a fixed installation to https://pypi.org/project/galax/ ! Should be resolved.
