### 2026-08-18 #10712 @eugvas
I confirm that the installation now proceeds smoothly; my only remaining suggestion for the paper is to add a reference to the Euclid paper in which this code was used.
