### 2026-08-23 0c01891
docs(paper): cite Euclid Q1 and Stream Legacy Survey papers
Add bib entries for Starkman et al. (2026, Euclid Q1) and Wu et al. (2026,
Potamides) and reference them where the paper describes ongoing research
using the package. Also add the missing citation for JAX-based splines.

Signed-off-by: nstarman <nstarman@users.noreply.github.com>
