+++ paper/paper.bib
@@ -99,6 +99,23 @@
   year    = {2017}
 }
 
+@article{Starkman:2026:EuclidQ1,
+  adsnote       = {Provided by the SAO/NASA Astrophysics Data System},
+  adsurl        = {https://ui.adsabs.harvard.edu/abs/2026arXiv260621774E},
+  archiveprefix = {arXiv},
+  author        = {{Euclid Collaboration} and {Starkman}, N. and {Nibauer}, J. and {Pearson}, S. and {Wu}, S. and {Walmsley}, M. and {Necib}, L. and {Bovy}, J. and {Marleau}, F.~R. and {Sola}, E. and {Scott}, D. and {Altieri}, B. and {Andreon}, S. and {Auricchio}, N. and {Baccigalupi}, C. and {Baldi}, M. and {Balestra}, A. and {Bardelli}, S. and {Battaglia}, P. and {Biviano}, A. and {Bolzonella}, M. and {Branchini}, E. and {Brescia}, M. and {Brinchmann}, J. and {Camera}, S. and {Capobianco}, V. and {Carbone}, C. and {Carretero}, J. and {Casas}, S. and {Castellano}, M. and {Castignani}, G. and {Cavuoti}, S. and {Chambers}, K.~C. and {Cimatti}, A. and {Colodro-Conde}, C. and {Congedo}, G. and {Conselice}, C.~J. and {Conversi}, L. and {Copin}, Y. and {Costille}, A. and {Courbin}, F. and {Courtois}, H.~M. and {Cropper}, M. and {Degaudenzi}, H. and {De Lucia}, G. and {Dole}, H. and {Dubath}, F. and {Duncan}, C.~A.~J. and {Dupac}, X. and {Dusini}, S. and {Escoffier}, S. and {Fabricius}, M. and {Farina}, M. and {Farinelli}, R. and {Ferriol}, S. and {Finelli}, F. and {Fosalba}, P. and {Fotopoulou}, S. and {Frailis}, M. and {Franceschi}, E. and {Fumana}, M. and {Gabarra}, L. and {Galeotta}, S. and {George}, K. and {Gillis}, B. and {Giocoli}, C. and {G{\'o}mez-Alvarez}, P. and {Gracia-Carpio}, J. and {Grazian}, A. and {Grupp}, F. and {Haugan}, S.~V.~H. and {Holmes}, W. and {Hormuth}, F. and {Hornstrup}, A. and {Huertas-Company}, M. and {Jahnke}, K. and {Jhabvala}, M. and {Joachimi}, B. and {Kermiche}, S. and {Kiessling}, A. and {Kubik}, B. and {K{\"u}mmel}, M. and {Kunz}, M. and {Kurki-Suonio}, H. and {Le Brun}, A.~M.~C. and {Ligori}, S. and {Lilje}, P.~B. and {Lindholm}, V. and {Lloro}, I. and {Magliocchetti}, M. and {Mainetti}, G. and {Mansutti}, O. and {Marggraf}, O. and {Martinelli}, M. and {Martinet}, N. and {Marulli}, F. and {Massey}, R.~J. and {Mauri}, N. and {Medinaceli}, E. and {Mei}, S. and {Melchior}, M. and {Meneghetti}, M. and {Merlin}, E. and {Meylan}, G. and {Monaco}, P. and {Mora}, A. and {Moresco}, M. and {Moretti}, C. and {Moscardini}, L. and {Nakajima}, R. and {Neissner}, C. and {Nichol}, R.~C. and {Niemi}, S.-M. and {Nightingale}, J.~W. and {Padilla}, C. and {Paltani}, S. and {Pasian}, F. and {Pedersen}, K. and {Percival}
[…]
