### 2026-08-23 6d7d452
chore: switch from pre-commit to prek
Co-authored-by: nstarman <8949649+nstarman@users.noreply.github.com>
