+++ noxfile.py
@@ -19,14 +19,14 @@ DIR = Path(__file__).parent.resolve()
 @session(uv_groups=["lint"], reuse_venv=True)
 def lint(s: nox.Session, /) -> None:
     """Run the linter."""
-    precommit(s)  # reuse pre-commit session
+    precommit(s)  # reuse prek session
     pylint(s)  # reuse pylint session
 
 
 @session(uv_groups=["lint"], reuse_venv=True)
 def precommit(s: nox.Session, /) -> None:
-    """Run pre-commit."""
-    s.run("pre-commit", "run", "--all-files", *s.posargs)
+    """Run prek."""
+    s.run("prek", "run", "--all-files", *s.posargs)
 
 
 @session(uv_groups=["lint"], reuse_venv=True)
