+++ pyproject.toml
@@ -66,7 +66,7 @@ docs = [
   "sphinx_design>=0.5.0",
 ]
 lint = [
-  "pre-commit>=4.5.0",
+  "prek>=0.4.0",
   "pylint==4.0.4",
 ]
 nox = [
