### 2026-08-23 5a5aeaf
style: prettier format CONTRIBUTING.md
Co-authored-by: nstarman <8949649+nstarman@users.noreply.github.com>
