### 2026-08-23 c80f59b
refactor: rename precommit nox session to prek_session and fix prek install docs
Co-authored-by: nstarman <8949649+nstarman@users.noreply.github.com>
