### 2026-08-23 6c31af9
Apply suggestions from code review
Co-authored-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
