### 2026-08-29 #10712 @TobiBu
@nstarman @augustfly could you please create a new release and point me to the zenodo archive so I can cross check all metadata?
