### 2026-09-01 011c509
Apply suggestion from @nstarman
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
