### 2026-09-01 6495e6d
Apply suggestion from @nstarman
Signed-off-by: Nathaniel Starkman <nstarman@users.noreply.github.com>
