### 2026-09-01 43c6b10
Fix prettier formatting in CONTRIBUTING.md
Co-authored-by: nstarman <8949649+nstarman@users.noreply.github.com>
