+++ CONTRIBUTING.md
@@ -56,8 +56,8 @@ management and [`nox`](https://nox.thea.codes/) to run common tasks.
    This installs potamides in editable mode along with the `dev` dependency
    group (tests, linters, docs, and nox).
 
-3. **Install the pre-commit hooks** so formatting and linting run automatically on
-   each commit:
+3. **Install the pre-commit hooks** so formatting and linting run automatically
+   on each commit:
 
    ```bash
    uv run prek install
