### 2026-09-01 #10712 @nstarman
@TobiBu I cut a new release and have Zenodo synced up — [10.5281/zenodo.17808542](https://doi.org/10.5281/zenodo.17808542)
