+++ .github/workflows/ossar.yml
@@ -50,6 +50,6 @@ jobs:
 
         # Upload results to the Security tab
       - name: Upload OSSAR results
-        uses: github/codeql-action/upload-sarif@v4.37.3
+        uses: github/codeql-action/upload-sarif@v4.37.9
         with:
           sarif_file: ${{ steps.ossar.outputs.sarifFile }}
