### 2026-09-05 #10712 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
